code/agents/learn-cc/requirements.txt
pyyaml>=6.0
python-dotenv>=1.0.0
litellm==1.82.6
